Major changes to the way names are resolved
Every type that can be resolved now extends Resolvable. This obsoletes
ScopedResult and also removes a lot of duplicated code.
Also, all of the conceptual type hierarchy has been filled in, and
there is now code in ASTConverter and NameResolver to convert types.
RelayedTypeParameter has been replaced by TypeArgumentInstance (a
TypeInstance which points to a TypeArgument).