Rework PointerTypes and remove TypeInstances
This allows more fine-grained control of types, and removes the need for
run-time checking that we have the correct type of PointerType. Instead, we can
now just use ClassPointerType, InterfacePointerType, etc. where before a
PointerType with a different TypeInstance would be used.

Also, ClassPointerType represents an inner class, and an OuterClassPointerType
is a special case of it where there is only one ConceptualClass/TypeParameter[]
combination. This makes an inner class more of a first-class citizen.