Resolve the universal base type properly
Previously, the AutomaticBaseClassPointerType was just a ClassPointerType with
no type information. This broke when things tried to find out which class it
was for, and got an ArrayIndexOutOfBoundsException.
Instead, this patch resolves the universal base type first, into a valid
OuterClassPointerType, which can then be used as normal.

The only special case this requires is to make sure that the base class of the
universal base class is always null, and that the code deals with this null
properly.