Add Type Parameter Bounds resolution
This commit also modifies the way PointerTypes are resolved, as type parameters
can now be used in place of other types, and also because the contents of type
argument lists must be checked.
One breakage described inside is in the resolution of type arguments, which is
not yet reliable for certain type hierarchies.