Add a check for invalid inner class inheritance
To extend an inner class (which is declared inside another class and is not
static), a child class must have an outer class which derives from the outer
class of the parent (or it can be contained in the same class as the parent).

This may need altering to allow for inner classes inside enums.