Refactor TypeResolver and fix parent resolution
The parent type resolution bug was caused by inner type definitions.
If an inner class of an enum extended the inner class of the enum's parent
class, it would not previously be resolved because class parents would always
be resolved before interface parents, and so type parameters and access
specifiers could not be checked.