Add the line text to ParseInfo; refactor errors
This removes the startLine and endLine fields of ParseInfo, and replaces them
with a simple line field. It also adds lineText, which contains the string that
was read in from the tokenizer.
This also moves the code to generate the location information into ParseInfo,
and adds a way of getting the highlighted line text from a ParseInfo.